# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "JavInfoApi",
    "description": "媒体元数据查询 API，基于 r18.dev 数据库（PostgreSQL）",
    "version": "1.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "本地开发服务器"
    }
  ],
  "paths": {
    "/api/v1/videos/search": {
      "get": {
        "summary": "搜索视频",
        "description": "多条件组合搜索视频。release_date_from/release_date_to 必须为 YYYY-MM-DD 格式，否则返回 400",
        "operationId": "searchVideos",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "关键词搜索（匹配 title_en、title_ja、comment_en、comment_ja）",
            "schema": { "type": "string" }
          },
          {
            "name": "content_id",
            "in": "query",
            "description": "内容ID精确匹配",
            "schema": { "type": "string" }
          },
          {
            "name": "dvd_id",
            "in": "query",
            "description": "DVD编号（自动处理横杠，不区分大小写）",
            "schema": { "type": "string" }
          },
          {
            "name": "maker_id",
            "in": "query",
            "description": "厂商ID（精确匹配，优先于 maker_name）",
            "schema": { "type": "integer" }
          },
          {
            "name": "maker_name",
            "in": "query",
            "description": "厂商名称（模糊匹配 name_en/name_ja）",
            "schema": { "type": "string" }
          },
          {
            "name": "series_id",
            "in": "query",
            "description": "系列ID（精确匹配，优先于 series_name）",
            "schema": { "type": "integer" }
          },
          {
            "name": "series_name",
            "in": "query",
            "description": "系列名称（模糊匹配 name_en/name_ja）",
            "schema": { "type": "string" }
          },
          {
            "name": "actress_id",
            "in": "query",
            "description": "演员ID（精确匹配，优先于 actress_name）",
            "schema": { "type": "integer" }
          },
          {
            "name": "actress_name",
            "in": "query",
            "description": "演员名称（模糊匹配 name_romaji/name_kanji/name_kana）",
            "schema": { "type": "string" }
          },
          {
            "name": "category_id",
            "in": "query",
            "description": "分类ID（精确匹配，优先于 category_name）",
            "schema": { "type": "integer" }
          },
          {
            "name": "category_name",
            "in": "query",
            "description": "分类名称（模糊匹配 name_en/name_ja）",
            "schema": { "type": "string" }
          },
          {
            "name": "label_id",
            "in": "query",
            "description": "品牌ID（精确匹配，优先于 label_name）",
            "schema": { "type": "integer" }
          },
          {
            "name": "label_name",
            "in": "query",
            "description": "品牌名称（模糊匹配 name_en/name_ja）",
            "schema": { "type": "string" }
          },
          {
            "name": "site_id",
            "in": "query",
            "description": "站点ID（1=DMM.com, 2=FANZA）",
            "schema": { "type": "integer" }
          },
          {
            "name": "year",
            "in": "query",
            "description": "发行年份（精确匹配，优先于 year_from/year_to）",
            "schema": { "type": "integer" }
          },
          {
            "name": "year_from",
            "in": "query",
            "description": "发行年份起始（含）",
            "schema": { "type": "integer" }
          },
          {
            "name": "year_to",
            "in": "query",
            "description": "发行年份结束（含）",
            "schema": { "type": "integer" }
          },
          {
            "name": "runtime_min",
            "in": "query",
            "description": "最小时长（分钟）",
            "schema": { "type": "integer" }
          },
          {
            "name": "runtime_max",
            "in": "query",
            "description": "最大时长（分钟）",
            "schema": { "type": "integer" }
          },
          {
            "name": "release_date_from",
            "in": "query",
            "description": "发售日期起始（YYYY-MM-DD，含）",
            "schema": { "type": "string", "format": "date" }
          },
          {
            "name": "release_date_to",
            "in": "query",
            "description": "发售日期结束（YYYY-MM-DD，含）",
            "schema": { "type": "string", "format": "date" }
          },
          {
            "name": "service_code",
            "in": "query",
            "description": "服务代码筛选（digital/mono/rental/ebook），空值不过滤",
            "schema": { "type": "string" }
          },
          {
            "name": "page",
            "in": "query",
            "description": "页码，从1开始",
            "schema": { "type": "integer", "default": 1 }
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "每页数量",
            "schema": { "type": "integer", "default": 20, "maximum": 100 }
          },
          {
            "name": "sort_by",
            "in": "query",
            "description": "排序字段及方向，格式：field1:asc,field2:desc（release_date/content_id/dvd_id/title_en/title_ja/runtime_mins），默认ASC",
            "schema": { "type": "string" }
          },
          {
            "name": "random",
            "in": "query",
            "description": "随机返回（random=1），与sort_by互斥",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/VideoListResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/videos": {
      "get": {
        "summary": "视频列表",
        "description": "获取视频分页列表",
        "operationId": "listVideos",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": { "type": "integer", "default": 1 }
          },
          {
            "name": "page_size",
            "in": "query",
            "schema": { "type": "integer", "default": 20, "maximum": 100 }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/VideoListResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/videos/{content_id}": {
      "get": {
        "summary": "视频详情",
        "description": "获取单个视频的完整信息",
        "operationId": "getVideo",
        "parameters": [
          {
            "name": "content_id",
            "in": "path",
            "required": true,
            "description": "视频内容ID",
            "schema": { "type": "string" }
          },
          {
            "name": "service_code",
            "in": "query",
            "description": "服务代码筛选",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/VideoDetailResponse" }
              }
            }
          },
          "400": {
            "description": "content_id 为空",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": {
            "description": "视频不存在",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "VideoListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/VideoSummary" }
          },
          "page": { "type": "integer" },
          "page_size": { "type": "integer" },
          "total_count": { "type": "integer" },
          "total_pages": { "type": "integer" }
        }
      },
      "VideoSummary": {
        "type": "object",
        "properties": {
          "content_id": { "type": "string" },
          "dvd_id": { "type": "string" },
          "title_en": { "type": "string" },
          "title_ja": { "type": "string" },
          "runtime_mins": { "type": "integer" },
          "release_date": { "type": "string", "format": "date" },
          "jacket_thumb_url": { "type": "string" },
          "site_id": { "type": "integer" },
          "service_code": { "type": "string" }
        }
      },
      "VideoDetailResponse": {
        "type": "object",
        "properties": {
          "content_id": { "type": "string" },
          "dvd_id": { "type": "string" },
          "title_en": { "type": "string" },
          "title_ja": { "type": "string" },
          "comment_en": { "type": "string" },
          "comment_ja": { "type": "string" },
          "runtime_mins": { "type": "integer" },
          "release_date": { "type": "string", "format": "date" },
          "sample_url": { "type": "string" },
          "maker_id": { "type": "integer" },
          "label_id": { "type": "integer" },
          "series_id": { "type": "integer" },
          "jacket_full_url": { "type": "string" },
          "jacket_thumb_url": { "type": "string" },
          "gallery_thumb_first": { "type": "string" },
          "gallery_thumb_last": { "type": "string" },
          "site_id": { "type": "integer" },
          "service_code": { "type": "string" },
          "maker": { "$ref": "#/components/schemas/Maker" },
          "label": { "$ref": "#/components/schemas/Label" },
          "series": { "$ref": "#/components/schemas/Series" },
          "actresses": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ActressWithOrdinality" }
          },
          "categories": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Category" }
          }
        }
      },
...
      "ActressWithOrdinality": {
        "type": "object",
        "description": "视频详情中的演员信息，包含排序序号",
        "properties": {
          "id": { "type": "integer" },
          "name_romaji": { "type": "string" },
          "name_kanji": { "type": "string" },
          "name_kana": { "type": "string" },
          "image_url": { "type": "string" },
          "ordinality": { "type": "integer", "description": "演员在该视频中的排序序号" }
        }
      },
      "Maker": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name_en": { "type": "string" },
          "name_ja": { "type": "string" }
        }
      },
...
      "Label": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name_en": { "type": "string" },
          "name_ja": { "type": "string" }
        }
      },
...
      "Series": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name_en": { "type": "string" },
          "name_ja": { "type": "string" }
        }
      },
...
      "Category": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name_en": { "type": "string" },
          "name_ja": { "type": "string" }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      }
    }
  }
}